remove predict function from training and added init prefect options
