+++ src/modelling/training.py
@@ -1,8 +1,9 @@
-import numpy as np
 import pandas as pd
 import xgboost as xgb
+from prefect import task
 
 
+@task("name=train-model", tags=["fails"], retries=3, retry_delay_seconds=60)
 def train_model(X_train: pd.DataFrame, y_train: pd.Series) -> xgb.XGBRegressor:
     """Given X_train and y_train, uses XGboost to train a model, returning the trained model.
 
@@ -21,22 +22,3 @@ def train_model(X_train: pd.DataFrame, y_train: pd.Series) -> xgb.XGBRegressor:
     model = xgb.XGBRegressor()
     model.fit(X_train, y_train)
     return model
-
-
-def predict(X_data: pd.DataFrame, model: xgb.XGBRegressor) -> np.ndarray:
-    """Given X_data and an XGBoost model, returns an array with the values predicted by the model.
-
-    Parameters
-    ----------
-    X_data: pd.DataFrame
-            Input data given to the model
-    model:  xgb.XGBRegressor
-            Model which will be used to predict the target variable of the input data.
-
-    Returns
-    -------
-    preds: np.ndarray
-           Numpy array containing the predicted values.
-    """
-    preds = model.predict(X_data)
-    return preds
